+++ ksa_compliance/ksa_compliance/doctype/zatca_phase_1_business_settings/zatca_phase_1_business_settings_list.js
@@ -1,0 +1,5 @@
+frappe.listview_settings["ZATCA Phase 1 Business Settings"] = {
+    onload(listview) {
+        ksa_compliance.premium_announcement.show_list_announcement_popup(listview)
+    }
+}